Reorganize clearExpandedTraceDefaultColors code
